backend/requirements.txt
django==5.2.5
django-ninja==1.4.3
simplejwt==2.0.1
pydantic==2.11.7
user-agents==2.2.0
channels==4.3.1
PyMySQL==1.1.1
django-redis==6.0.0
channels_redis==4.3.0
uvicorn==0.35.0
uvicorn[standard]==0.35.0
celery==5.5.3
openai==1.101.0
python-dotenv==1.1.1
torch==2.8.0
torchaudio==2.8.0
modelscope
ffmpeg-python
pydub
psutil
django-stubs==5.2.5
markdown==3.9
